Changed home page, updated profiles, manage.py... need to build out leagues functionality next
